Fix groundtype string log and argpp public include paths
string_to_type was casting a std::string to int for logging; log the
string itself. type_to_string still casts the GPType enum.

For argpp, expose both include/ and include/argpp/ as PUBLIC
BUILD_INTERFACE so consumers (#include <argpp/argpp.hpp>) and the
library tests (#include <argpp.hpp>) both compile. Keep
INSTALL_INTERFACE as include/ for the installed layout.

+++ src/pingus/groundtype.cpp
@@ -39,7 +39,7 @@ Groundtype::string_to_type(std::string const& arg_type)
     return Groundtype::GP_REMOVE;
   else
   {
-    log_error("Groundtype: Unhandled type: '{}'", static_cast<int>(arg_type));
+    log_error("Groundtype: Unhandled type: '{}'", arg_type);
     return Groundtype::GP_GROUND;
   }
 }
